+++ packages/jupyterlab-gridstack/src/editor/components/metadata.tsx
